Refactor utility functions into separate 'util' directory
