+++ src/glob-tracker.js
@@ -1,20 +1,4 @@
-/**
- * Determines if two arrays have the same values.
- *
- * Only applicable with arrays which contain basic types.
- */
-const arraysEqual = (a, b) => {
-  if (a.length !== b.length) {
-    return false;
-  }
-  const aExtra = a.some((value) => {
-    return !(b.includes(value));
-  });
-  const bExtra = b.some((value) => {
-    return !(a.includes(value));
-  });
-  return !(aExtra || bExtra);
-}
+const arraysEqual = require('./util/arrays-equal.js');
 
 /**
  * Tracks changes in glob results over a period of time.
